% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parties.R
\name{parties}
\alias{parties}
\title{List of parties up for election}
\source{
https://www.valg.im.dk/partier-og-kandidater/opstillingsberettigede-partiers-adresser-mv
}
\usage{
parties
}
\value{
Returns a tibble with information on the major parties up for election
}
\description{
List of parties up for election
}
\examples{
head(parties)
}
