% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_bib.R
\name{pkg_bib}
\alias{pkg_bib}
\title{Create bibliography of \code{{renv}}-installed packages}
\usage{
pkg_bib(dependencies = FALSE, filename = "renv_packages.bib")
}
\description{
This was inspired by https://github.com/rstudio/renv/issues/340
}
