% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nchar.R
\name{plot_nchar}
\alias{plot_nchar}
\title{Get character count from \code{ggplot2} object.}
\usage{
plot_nchar(plot, debug = FALSE)
}
\arguments{
\item{plot}{Input a \code{ggplot2} object or a \code{patchwork} object.}

\item{debug}{Logical. If \code{TRUE}, prints the text values that are counted.}
}
\value{
Returns an integer with the amount of characters in the plot.
Spaces are included.
}
\description{
Counts the number of characters in visible text elements from a \code{ggplot2}
plot. This also supports plots composed with \code{patchwork}, including collected
guides.
}
\examples{
plot <- data.frame(
  x = seq(1, 10),
  y = seq(10, 1),
  label = rep(seq(1, 10, 2), 2)
  ) |>
  ggplot2::ggplot(ggplot2::aes(x = x, y = y, label = label)) +
  ggplot2::geom_text(nudge_y = 3)

plot_nchar(plot)
}
