% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_alek_map.R
\name{theme_map}
\alias{theme_map}
\title{theme_alek_map()
A clean theme that is good for displaying maps from
\code{\link[ggplot2]{geom_map}()}.}
\usage{
theme_map(base_size = 9, base_family = "")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
theme_alek_map()
A clean theme that is good for displaying maps from
\code{\link[ggplot2]{geom_map}()}.
}
