% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_alek_map.R
\name{theme_map_color}
\alias{theme_map_color}
\title{A color-ready theme made for maps from \code{ggplot2::geom_map()}.}
\usage{
theme_map_color(
  color = "green",
  secondary_color = "pink",
  background_color = "black",
  base_size = 9,
  base_family = ""
)
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
This theme is made color-ready specifically for maps.
}
