% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_num.R
\name{format_num}
\alias{format_num}
\title{Format a number for presentation}
\usage{
format_num(
  x,
  decimals = 2,
  rounding = TRUE,
  big.mark = ".",
  decimal.mark = ","
)
}
\arguments{
\item{x}{An integer/double value}

\item{big.mark}{The character that divides numbers in thousands}

\item{decimal.mark}{The decimal separator}
}
\value{
A number formatted correctly with two (default) decimals, a comma as the decimal separator and a period as the thousand-separator.
}
\description{
Format a number for presentation
}
\examples{
x <- pi
format_num(x)

}
