% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parties.R
\docType{data}
\name{parties}
\alias{parties}
\title{List of parties up for election}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 12 rows and 4 columns.
}
\source{
https://www.valg.im.dk/partier-og-kandidater/opstillingsberettigede-partiers-adresser-mv
}
\usage{
parties
}
\value{
Returns a tibble with information on the major parties up for election
}
\description{
List of parties up for election
}
\examples{
head(parties)
}
\keyword{datasets}
