% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_bib.R
\name{pkg_bib}
\alias{pkg_bib}
\title{Create bibliography of \code{{renv}}-installed packages}
\usage{
pkg_bib(dependencies = FALSE, filename = "packages.bib", targets = FALSE)
}
\arguments{
\item{dependencies}{Boolean to define how many packages are cited.}

\item{filename}{Specify a filename/path for the \code{.bib}.}

\item{targets}{Boolean to define if the bibliography should be created based
on the packages defined in a targets pipeline and saved to
\verb{_targets_packages.R} by the function \code{targets::tar_renv()}}
}
\value{
Usually called for its side effect of a \code{.bib} written to the
provided paths. Also outputs the package names in the console.
}
\description{
This was inspired by https://github.com/rstudio/renv/issues/340
}
