% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_nchar.R
\name{plot_nchar}
\alias{plot_nchar}
\title{Get character count from \code{ggplot2} object.}
\usage{
plot_nchar(plot)
}
\arguments{
\item{plot}{Input a \code{ggplot2} object}
}
\value{
Returns a integer with the amount of characters in the plot
}
\description{
Get character count from \code{ggplot2} object.
}
\examples{
plot <- data.frame(
  x = seq(1, 10),
  y = seq(10, 1),
  label = rep(seq(1, 10, 2), 2)
  ) |>
  ggplot2::ggplot(ggplot2::aes(x = x, y = y, label = label)) +
  ggplot2::geom_text(nudge_y = 3)

plot_nchar(plot)
}
