% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_alek_map.R
\name{theme_map}
\alias{theme_map}
\title{A theme made for maps from  \code{ggplot2::geom_map()}.}
\usage{
theme_map(base_size = 9, base_family = "")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
This theme is made specifically for maps made with. A color-ready version exists as \code{theme_map_color()}.
}
